python code to solve day 14
This uses the Python pulp module, and is heavily influenced by

https://github.com/Recursing/AdventOfCode-2019/blob/master/day14_pulp.py

I made some changes to the parsing and data structures, but it's basically
their code that interfaced with pulp.

I found the pulp module but couldn't figure out how to get it to work.
This person (who contacted me on twitter) saw my tweet about it, figured
it out, and shared his code with me. So it's not really *cheating*,
it's teamwork!